feat: 支持动态注册功能、移除 iot_c_sdk 子模块并支持通过宏关闭 TLS

1. 支持动态注册功能

2. 移除 iot_c_sdk 子模块但保留其内容，并将其添加到 .gitignore

3. 支持 TLS，可以通过关闭宏 #define AUTH_WITH_NO_TLS 来实现

+++ components/qcloud_iot_c_sdk/iot_explorer_sdk/inc/utils_downloader.h
@@ -34,6 +34,7 @@ extern "C" {
 #endif
 
 #include <stdio.h>
+#include <stdint.h>
 
 /**
  * @brief Downloader status.
@@ -54,7 +55,7 @@ typedef enum {
  */
 typedef struct {
     // memory
-    void* (*downloader_malloc)(size_t len); /**< user malloc */
+    void* (*downloader_malloc)(uint32_t len); /**< user malloc */
     void (*downloader_free)(void* val);     /**< user free */
 
     // break point
